[FIX] o-multiverse: handle empty values in branch and repo arrays
When the user provides an empty value in the `--branches` option's list,
the script would crash when trying to create a directory for it.

The same happens when excluding repositories using the `--exclude`
option. The values are removed from the array, but they're still empty
values that we can loop over.

We now check each time whether we have an empty value or not.