[IMP] docker: Improved Docker configuration
Took a lot of inspiration from the [repo of Denis] to match our
production environment as closely as possible.

Also made it so to build the Docker image and run the container using
the amd64 architecture, so we can use Odoo's packages built for amd64
also on Apple Silicon chips using the Rosetta emulator.

[repo of Denis]: https://github.com/beledouxdenis/docker-odoo-ready